Work

+++ src/main/java/ch/qligier/heicofawallpaper/model/PhaseSolar.java
@@ -1,12 +1,14 @@
 package ch.qligier.heicofawallpaper.model;
 
 /**
- * HeicOfAWallpaper
+ * A phase of a solar dynamic wallpaper.
  *
  * @param frameIndex The index of the frame to show.
+ * @param elevation  The sun elevation.
+ * @param azimuth    The sun azimuth.
  * @author Quentin Ligier
  **/
-public record PhaseSolar(int frameIndex,
+public record PhaseSolar(short frameIndex,
                          float elevation,
                          float azimuth) {
 }
